Move matcher tests
